refactor : 베팅 금액과 승리 금액 타입 변경 (Long -> Double)

+++ src/main/kotlin/betting/Bet.kt
@@ -2,11 +2,11 @@ package betting
 
 @JvmInline
 value class Bet(
-    val amount: Long = 0L,
+    val amount: Double = 0.0,
 ) {
     init {
         require(amount >= 0) { "베팅 금액은 음수일 수 없습니다." }
     }
 
-    fun negative(): Long = amount * -1L
+    fun negative(): Double = amount * -1L
 }
